tests/wpt/tests/websockets/constructor/018.html

<!doctype html>
<title>WebSockets: NULL char in url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo-query?x\u0000y\u0000');
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, 'x%00y');
    ws.close();
    t.done();
  })
  ws.onclose = ws.onerror = t.step_func(function(e) {assert_unreached(e.type)});
});
</script>
